Try to avoid multiple texture uploads for YUV video
I believe that the sheer count of texture uploads is a cause of
slowdowns especially on Linux + Nvidia.

This is my attempt at sampling the "reshaped"

YUV420p texture were a single packed texture is returned, but the u and
v are downsample by 2 in both axes making it tricky to access the right
data.

Still very much a WIP